requirements.txt
Appium-Python-Client==0.24
python-dateutil==2.6.1
selenium==3.7.0
six==1.11.0
timeout-decorator==0.4.0
